Button Color 설정중,  ImageUI GIO 수정, 부모 자식 오브젝트 관계 설정 해지시 행렬연산중 (scale완료)

+++ Project/Engine/Function.cpp
@@ -32,7 +32,7 @@ void TStringToArr(const tstring& _in, char* _out, int _iSize)
 void StringTool::StringToLower(string& _str)
 {
 	int i = 0;
-	int iStrSize = _str.size();
+	int iStrSize = (int)_str.size();
 	while (i < iStrSize) {
 		if (std::isupper(_str[i])) {
 			_str[i] = std::tolower(_str[i]);
@@ -44,7 +44,7 @@ void StringTool::StringToLower(string& _str)
 void StringTool::StringToUpper(string& _str)
 {
 	int i = 0;
-	int iStrSize = _str.size();
+	int iStrSize = (int)_str.size();
 	while (i < iStrSize) {
 		if (std::islower(_str[i])) {
 			_str[i] = std::toupper(_str[i]);
